feat: add subheading support to tabs, initial docs
