+++ src/utils/markdown/plugins/tabs/index.ts
@@ -1,0 +1,1 @@
+export * from './tabs';
